Split CLI build graph loading
